+++ src/interpreter/env.rs
@@ -28,8 +28,6 @@ pub struct FuncKind {
     pub params: Vec<(String, String)>,
     pub is_async: bool,
     pub is_pub: bool,
-    pub is_impl: bool,
-    pub is_mut: bool,
 }
 
 #[derive(Clone, Debug, PartialEq)]
@@ -133,7 +131,7 @@ impl Env {
         ));
     }
 
-    pub fn _define_mod_func(&self, source: String, f: LiteralType, k: String, v: FuncKind) {
+    pub fn define_mod_func(&self, source: String, f: LiteralType, k: String, v: FuncKind) {
         let mut mod_vals = self.mod_vals.borrow_mut();
         let entry = mod_vals.entry(source).or_insert_with(Vec::new);
         entry.push((